Lab 12 - Add a Chart
Added a chart using Chart.js that displays the number of votes and
views for each product after a round of votes.

Also modified the behavior so that the images rendered do not include
any from the previous round.

Modified to better handle displaying an arbitrary number of images
per round.